+++ src/main/java/spark/Greetable.java
@@ -1,5 +1,0 @@
-package spark;
-
-public interface Greetable {
-    public String greet(String name);
-}