+++ src/main/java/spark/HelloGreeter.java
@@ -1,8 +1,0 @@
-package spark;
-
-public class HelloGreeter implements Greetable{
-    @Override
-    public String greet(String name) {
-        return "Hello, " + name;
-    }
-}