+++ src/main/java/spark/ProjectServlet.java
@@ -7,17 +7,14 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-//import org.apache.spark.SparkConf;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.api.java.JavaPairRDD;
 import org.apache.spark.api.java.JavaRDD;
 
-
-import scala.Tuple2;
-
 public class ProjectServlet extends HttpServlet {
 
     JavaSparkContext sparkContext;
+    JavaRDD<String> inputFile;
 
     private static final long serialVersionUID = 1L;
 
@@ -25,40 +22,12 @@ public class ProjectServlet extends HttpServlet {
 
     }
 
-    public ProjectServlet(JavaSparkContext context)
+    public ProjectServlet(JavaSparkContext context, JavaRDD<String> rdd)
     {
         this.sparkContext = context;
+        this.inputFile = rdd;
     }
 
-    public JavaRDD<String> rddFiltering(JavaRDD<String> rdd, String text, int column) {
-        
-        JavaRDD<String> rdd2 = rdd.filter(x -> {
-            String[] splitRow = x.split(",");
-  
-            return splitRow[column].equals(text);
-        });
-        return rdd2;
-    }
-
-    public JavaRDD<String> rddStripToColumn(JavaRDD<String> rdd, int column)
-    {
-        JavaRDD<String> rdd2 = rdd.map(x ->
-        {
-            String[] splitRow = x.split(",");
-            return splitRow[column];
-        });
-        return rdd2;
-    }
-   
-    public JavaPairRDD<String, Integer> rddCounterString(JavaRDD<String> rdd) {
-        JavaPairRDD<String, Integer> rdd2 = rdd.mapToPair((x) -> new Tuple2<>(x, 1));
-        return rdd2.reduceByKey((x, y) -> (x + y));
-    }
-
-    public JavaPairRDD<Integer, Integer> rddCounterInteger(JavaRDD<String> rdd) {
-        JavaPairRDD<Integer, Integer> rdd2 = rdd.mapToPair(x -> new Tuple2<>(Integer.parseInt(x), 1));
-        return rdd2.reduceByKey((x, y) -> (x + y));
-    }
 
     public int numberOfCasesInYear(JavaRDD<String> rdd, int year)
     {
@@ -92,32 +61,168 @@ public class ProjectServlet extends HttpServlet {
             return "";
         });
 
-        JavaPairRDD<String, Integer> rdd4 = rddCounterString(rdd3);
+        JavaPairRDD<String, Integer> rdd4 = RDDCustomOperations.rddCounterString(rdd3);
 
         return rdd4.lookup(String.valueOf(year)).get(0);         
     
     }
 
-    public JavaPairRDD<String,Integer> listOfCasesByYear(JavaRDD<String> rdd)
+
+
+    // public JavaPairRDD<String,Integer> sightingsByYear(JavaRDD<String> rdd)
+    // {
+    //     JavaRDD<String> rdd2 = rdd.map(x -> {
+    //         try {
+    //             String[] splitRow = x.split(",");
+    //             String[] splitRow2 = splitRow[0].split(" ");
+    //             String[] splitRow3 = splitRow2[0].split("/");
+    //             return splitRow3[2];
+    //         } catch (Exception e) {
+    //             System.out.println("Exception inside rdd2 of listOfCasesByYear");
+    //             e.printStackTrace();
+    //         }
+    //         return "";
+    //     });
+
+    //     JavaPairRDD<String, Integer> rdd3 = RDDCustomOperations.rddCounterString(rdd2);
+    //     return rdd3;
+  
+    // }
+    
+    // public JavaPairRDD<String,Integer> sightingsByHour(JavaRDD<String> rdd)
+    // {
+    //     JavaRDD<String> rdd2 = rdd.map(x -> {
+    //         try {
+    //             String[] splitRow = x.split(",");
+    //             String[] splitRow2 = splitRow[0].split(" ");
+    //             String[] splitRow3 = splitRow2[1].split(":");
+    //             return splitRow3[0];
+    //         } catch (Exception e) {
+    //             System.out.println("Exception inside rdd2 of listOfCasesByYear");
+    //             e.printStackTrace();
+    //         }
+    //         return "";});
+
+    //         JavaPairRDD<String, Integer> rdd3 = RDDCustomOperations.rddCounterString(rdd2);
+    //         return rdd3;
+    // }
+
+    // public JavaPairRDD<String,Integer> sightingsByMonth(JavaRDD<String> rdd)
+    // {
+    //     JavaRDD<String> rdd2 = rdd.map(x -> {
+    //         try {
+    //             String[] splitRow = x.split(",");
+    //             String[] splitRow2 = splitRow[0].split(" ");
+    //             String[] splitRow3 = splitRow2[0].split("/");
+    //             return splitRow3[0];
+    //         } catch (Exception e) {
+    //             System.out.println("Exception inside rdd2 of listOfCasesByYear");
+    //             e.printStackTrace();
+    //         }
+    //         return "";
+    //     });
+
+    //     JavaRDD<String> rdd3 = rdd2.filter(x -> {
+
+    //         return !x.equals("datetime");
+    //     });
+
+
+    //     JavaPairRDD<String, Integer> rdd4 = RDDCustomOperations.rddCounterString(rdd3);
+
+    //     return rdd4;
+    // }
+    
+
+
+    // public JavaPairRDD<String,Integer> sightingsInState(JavaRDD<String> rdd,String state)
+    // {
+    //     JavaRDD<String> rdd2 = RDDCustomOperations.rddFiltering(rdd,state.toLowerCase(),2);
+
+    //     JavaRDD<String> rdd3 = rdd2.map(x ->
+    //     {
+    //         String[] splitRow = x.split(",");
+    //         return splitRow[1] + "," + splitRow[2];
+    //     });
+
+    //     return RDDCustomOperations.rddCounterString(rdd3);
+    // }
+
+    // public int numberOfCasesInState(JavaRDD<String> rdd,String state)
+    // {
+    //     JavaRDD<String> rdd2 = RDDCustomOperations.rddFiltering(rdd,state.toLowerCase(),2);
+    //     JavaRDD<String> rdd3 = RDDCustomOperations.rddStripToColumn(rdd2,2);
+    //     JavaPairRDD<String,Integer> rdd4 = RDDCustomOperations.rddCounterString(rdd3);
+ 
+    //     return rdd4.lookup(state.toLowerCase()).get(0);
+    // }
+
+
+    // public JavaPairRDD<String,Integer> sightingsInCountry(JavaRDD<String> rdd,String country)
+    // {
+    //     JavaRDD<String> rdd2 = RDDCustomOperations.rddFiltering(rdd,country.toLowerCase(),3);
+
+    //     JavaRDD<String> rdd3 = rdd2.map(x ->
+    //     {
+    //         String[] splitRow = x.split(",");
+    //         return splitRow[2]+","+splitRow[3];
+    //     });
+
+    //     return RDDCustomOperations.rddCounterString(rdd3);
+    // }
+
+    // public int numberOfCasesInCountry(JavaRDD<String> rdd,String country)
+    // {
+    //     JavaRDD<String> rdd2 = RDDCustomOperations.rddFiltering(rdd,country.toLowerCase(),3);
+    //     JavaRDD<String> rdd3 = RDDCustomOperations.rddStripToColumn(rdd2,3);
+    //     JavaPairRDD<String,Integer> rdd4 = RDDCustomOperations.rddCounterString(rdd3);
+
+    //     return rdd4.lookup(country.toLowerCase()).get(0);
+    // }
+
+
+    public JavaPairRDD<String,Integer> sightingsByShape(JavaRDD<String> rdd)
+    {
+        JavaRDD<String> rdd2 = RDDCustomOperations.rddStripToColumn(rdd,4);
+        JavaPairRDD<String,Integer> rdd3 = RDDCustomOperations.rddCounterString(rdd2);
+
+        return rdd3;
+    }
+    
+
+    public JavaPairRDD<String,Integer> listOfAllCases (JavaRDD<String> rdd)
+    {
+        JavaRDD<String> rdd2 = rdd.map(x -> {
+            String[] splitRow = x.split(",");
+  
+            return splitRow[2] + "," + splitRow[3];
+        });
+
+        JavaPairRDD<String,Integer> rdd3 = RDDCustomOperations.rddCounterString(rdd2);
+ 
+        return rdd3;
+    }
+    
+
+
+    public JavaPairRDD<String,Integer> listByDatesPosted(JavaRDD<String> rdd)
     {
         JavaRDD<String> rdd2 = rdd.map(x -> {
             try {
                 String[] splitRow = x.split(",");
-                String[] splitRow2 = splitRow[0].split(" ");
-                String[] splitRow3 = splitRow2[0].split("/");
-                return splitRow3[2];
+                String[] splitRow2 = splitRow[8].split("/");
+                return splitRow2[2];
             } catch (Exception e) {
-                System.out.println("Exception inside rdd2 of listOfCasesByYear");
+                System.out.println("Exception inside rdd2 of listOfDatesPosted");
                 e.printStackTrace();
             }
             return "";
         });
 
-        JavaPairRDD<String, Integer> rdd3 = rddCounterString(rdd2);
-        return rdd3;
-  
+        return RDDCustomOperations.rddCounterString(rdd2);
     }
-   
+
+
     public double computePercentageChange(JavaRDD<String> rdd,int year1, int year2)
     { 
         //Returns the rows that have year1 or year2
@@ -151,7 +256,7 @@ public class ProjectServlet extends HttpServlet {
             return "";
         });
 
-        JavaPairRDD<String, Integer> rdd4 = rddCounterString(rdd3);
+        JavaPairRDD<String, Integer> rdd4 = RDDCustomOperations.rddCounterString(rdd3);
 
         double firstKey = rdd4.lookup(String.valueOf(year1)).get(0);
         double secondKey = rdd4.lookup(String.valueOf(year2)).get(0);
@@ -159,112 +264,21 @@ public class ProjectServlet extends HttpServlet {
         return (secondKey -  firstKey) /  firstKey * 100;
     }
 
-    public JavaPairRDD<String,Integer> listOfCasesInState(JavaRDD<String> rdd,String country)
-    {
-        JavaRDD<String> rdd2 = rddFiltering(rdd,country.toLowerCase(),2);
 
-        JavaRDD<String> rdd3 = rdd2.map(x ->
-        {
-            String[] splitRow = x.split(",");
-            return splitRow[1] + "," + splitRow[2];
-        });
+   
 
-        return rddCounterString(rdd3);
-    }
+   
 
-    public int numberOfCasesInState(JavaRDD<String> rdd,String state)
-    {
-        JavaRDD<String> rdd2 = rddFiltering(rdd,state.toLowerCase(),2);
-        JavaRDD<String> rdd3 = rddStripToColumn(rdd2,2);
-        JavaPairRDD<String,Integer> rdd4 = rddCounterString(rdd3);
- 
-        return rdd4.lookup(state.toLowerCase()).get(0);
-    }
-
-    public JavaPairRDD<String,Integer> listOfAllCases (JavaRDD<String> rdd)
-    {
-        JavaRDD<String> rdd2 = rdd.map(x -> {
-            String[] splitRow = x.split(",");
-  
-            return splitRow[2] + "," + splitRow[3];
-        });
-
-        JavaPairRDD<String,Integer> rdd3 = rddCounterString(rdd2);
- 
-        return rdd3;
-    }
     
-    public JavaPairRDD<String,Integer> listOfCasesInCountry(JavaRDD<String> rdd,String country)
-    {
-        JavaRDD<String> rdd2 = rddFiltering(rdd,country.toLowerCase(),3);
 
-        JavaRDD<String> rdd3 = rdd2.map(x ->
-        {
-            String[] splitRow = x.split(",");
-            return splitRow[2]+","+splitRow[3];
-        });
-
-        return rddCounterString(rdd3);
-    }
-
-    public int numberOfCasesInCountry(JavaRDD<String> rdd,String country)
-    {
-        JavaRDD<String> rdd2 = rddFiltering(rdd,country.toLowerCase(),3);
-        JavaRDD<String> rdd3 = rddStripToColumn(rdd2,3);
-        JavaPairRDD<String,Integer> rdd4 = rddCounterString(rdd3);
-
-        return rdd4.lookup(country.toLowerCase()).get(0);
-    }
-
-    //Get rid of (datetime,1) somehow
-    public JavaPairRDD<String,Integer> sightingsByMonth(JavaRDD<String> rdd)
-    {
-        JavaRDD<String> rdd2 = rdd.map(x -> {
-            try {
-                String[] splitRow = x.split(",");
-                String[] splitRow2 = splitRow[0].split(" ");
-                String[] splitRow3 = splitRow2[0].split("/");
-                return splitRow3[0];
-            } catch (Exception e) {
-                System.out.println("Exception inside rdd2 of listOfCasesByYear");
-                e.printStackTrace();
-            }
-            return "";
-        });
-
-        JavaRDD<String> rdd3 = rdd2.filter(x -> {
-
-            return !x.equals("datetime");
-        });
-
-
-        JavaPairRDD<String, Integer> rdd4 = rddCounterString(rdd3);
-
-        return rdd4;
-    }
-
-    public JavaPairRDD<String,Integer> listByDatesPosted(JavaRDD<String> rdd)
-    {
-        JavaRDD<String> rdd2 = rdd.map(x -> {
-            try {
-                String[] splitRow = x.split(",");
-                String[] splitRow2 = splitRow[8].split("/");
-                return splitRow2[2];
-            } catch (Exception e) {
-                System.out.println("Exception inside rdd2 of listOfDatesPosted");
-                e.printStackTrace();
-            }
-            return "";
-        });
-
-        return rddCounterString(rdd2);
-    }
+    
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         
-        JavaRDD<String> rdd = sparkContext
-                 .textFile("C:\\Users\\Garrison\\Project-1-Garrison\\src\\main\\resources\\scrubbed.csv", 4);
+        resp.getWriter().println("In project");
+
+        //JavaRDD<String> rdd = sparkContext.textFile("C:\\Users\\Garrison\\Project-1-Garrison\\src\\main\\resources\\scrubbed.csv", 4);
 
         // double change = computePercentageChange(rdd,2000,2005);
         // resp.getWriter().println("Percentage Change between 2000 and 2005");
@@ -284,16 +298,22 @@ public class ProjectServlet extends HttpServlet {
         // resp.getWriter().println("List of Cases by year");
         // resp.getWriter().println(listOfCasesByYear(rdd).sortByKey().collect());
 
-        resp.getWriter().println("List of Cases in year 2013");
-        resp.getWriter().println(numberOfCasesInYear(rdd, (int) 2013));
+        // resp.getWriter().println("List of Cases in year 2013");
+        // resp.getWriter().println(numberOfCasesInYear(rdd, (int) 2013));
 
-        resp.getWriter().println("List of Dates Posted");
-        resp.getWriter().println(listByDatesPosted(rdd).sortByKey().collect());
+        // resp.getWriter().println("List of Dates Posted");
+        // resp.getWriter().println(listByDatesPosted(rdd).sortByKey().collect());
 
-        resp.getWriter().println("Sightings By Month");
-        resp.getWriter().println(sightingsByMonth(rdd).sortByKey().collect());
-        resp.getWriter().println(sightingsByMonth(rdd).sortByKey(true).collect());
-        resp.getWriter().println(sightingsByMonth(rdd).sortByKey(false).collect());
+        // resp.getWriter().println("Sightings By Month");
+        // resp.getWriter().println(sightingsByMonth(rdd).sortByKey().collect());
+        // resp.getWriter().println(sightingsByMonth(rdd).sortByKey(true).collect());
+        // resp.getWriter().println(sightingsByMonth(rdd).sortByKey(false).collect());
+
+        // resp.getWriter().println("List of Cases by shape");
+        // resp.getWriter().println(listOfCasesByShape(rdd).sortByKey().collect());
+
+        // resp.getWriter().println("List of Cases by hour");
+        // resp.getWriter().println(listOfCasesByHour(rdd).sortByKey().collect());
 
     
     }