+++ src/main/java/spark/Server.java
@@ -3,6 +3,7 @@ package spark;
 import java.io.File;
 
 import org.apache.spark.SparkConf;
+import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.catalina.LifecycleException;
 import org.apache.catalina.Wrapper;
@@ -13,15 +14,26 @@ public class Server {
     public static void main(String[] args) throws LifecycleException {
 
         SparkConf conf = new SparkConf().setAppName("Project1").setMaster("local[8]");
+        
         JavaSparkContext sparkContext = new JavaSparkContext(conf);
+
+        JavaRDD<String> rdd = sparkContext.textFile("C:\\Users\\Garrison\\Project-1-Garrison\\src\\main\\resources\\scrubbed.csv", 4);
 
         Tomcat tomcat = new Tomcat();
         tomcat.setBaseDir(new File("target/tomcat/").getAbsolutePath());
         tomcat.setPort(8080);
         tomcat.getConnector();
         tomcat.addWebapp("/spark", new File("src/main/").getAbsolutePath());
-        Wrapper projectServlet = tomcat.addServlet("/spark", "ProjectServlet", new ProjectServlet(sparkContext));
+        Wrapper projectServlet = tomcat.addServlet("/spark", "ProjectServlet", new ProjectServlet(sparkContext,rdd));
         projectServlet.addMapping("/project");
+        Wrapper byTimeServlet = tomcat.addServlet("/spark", "ByTimeServlet", new ByTimeServlet(sparkContext,rdd));
+        byTimeServlet.addMapping("/time");
+        Wrapper byStateServlet = tomcat.addServlet("/spark", "ByStateServlet", new ByTimeServlet(sparkContext,rdd));
+        byStateServlet.addMapping("/state");
+        Wrapper byCountryServlet = tomcat.addServlet("/spark", "ByCountryServlet", new ByCountryServlet(sparkContext,rdd));
+        byCountryServlet.addMapping("/country");
+       
+
         tomcat.start();
 
         Runtime.getRuntime().addShutdownHook(new Thread() {